Change counter type to Long

+++ frontend/compose-app/data/network/profile/api/src/main/java/cc/snakechat/data/profile/ProfileResponse.kt
@@ -5,14 +5,14 @@ import kotlinx.serialization.Serializable
 @Serializable
 class ProfileResponse(
     @SerialName("followers_count")
-    val followersCount: Int? = null,
+    val followersCount: Long? = null,
     @SerialName("following_count")
-    val followingCount: Int? = null,
+    val followingCount: Long? = null,
     val id: String? = null,
     @SerialName("post_thumbnails")
     val postThumbnails: List<PostThumbnail>? = null,
     @SerialName("posts_count")
-    val postsCount: Int? = null,
+    val postsCount: Long? = null,
     @SerialName("profile_url")
     val profileUrl: String? = null,
     val status: String? = null,
