+++ frontend/compose-app/domain/profile/src/main/kotlin/cc/snakechat/domain/profile/ProfileResponse.kt
@@ -1,11 +1,11 @@
 package cc.snakechat.domain.profile
 
 class Profile(
-    val followersCount: Int = 0,
-    val followingCount: Int = 0,
+    val followersCount: Long = 0,
+    val followingCount: Long = 0,
     val id: String,
     val postThumbnails: List<PostThumbnail> = emptyList(),
-    val postsCount: Int = 0,
+    val postsCount: Long = 0,
     val profileUrl: String? = null,
     val status: String? = null,
     val username: String
